Move map use cases into their own folder
