+++ src/test/java/com/mg105/use_cases/CharacterMoverTest.java
@@ -2,6 +2,8 @@ package com.mg105.use_cases;
 
 import com.mg105.entities.*;
 import com.mg105.entities.items.UpgradeToken;
+import com.mg105.use_cases.map.CharacterMover;
+import com.mg105.use_cases.map.RoomUpdater;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
